workspace-fox-smarthome/skills/robot-car/references/raspberry-pi-code/requirements.txt
# Robot Car 依赖列表

# Web 框架
flask==3.0.0
flask-cors==4.0.0

# 视觉处理
opencv-python==4.8.1.78
numpy==1.24.3

# 硬件控制
RPi.GPIO==0.7.1
adafruit-circuitpython-servokit==1.3.18
adafruit-circuitpython-motorkit==1.6.13

# 音频播放
pygame==2.5.2

# 工具库
python-dotenv==1.0.0
